requirements.txt
numpy
pillow
tk
decord
opencv-python
tqdm